refactor Transit to interface
