+++ java/test/jmri/TransitTest.java
@@ -5,6 +5,8 @@ import jmri.util.JUnitUtil;
 import org.junit.jupiter.api.*;
 import org.junit.Assert;
 
+import jmri.implementation.DefaultTransit;
+
 /**
  * Tests for Transit class.
  *
@@ -16,12 +18,12 @@ public class TransitTest {
 
    @Test
    public void SysNameConstructorTest(){
-      Assert.assertNotNull("Constructor", new Transit("TT1"));
+      Assert.assertNotNull("Constructor", new DefaultTransit("TT1"));
    }
 
    @Test
    public void TwoNameStringConstructorTest(){
-      Assert.assertNotNull("Constructor", new Transit("TT1", "user name"));
+      Assert.assertNotNull("Constructor", new DefaultTransit("TT1", "user name"));
    }
 
    @BeforeEach
