+++ java/test/jmri/jmrit/dispatcher/ActiveTrainTest.java
@@ -13,7 +13,7 @@ public class ActiveTrainTest {
 
     @Test
     public void testCTor() {
-        jmri.Transit transit = new jmri.Transit("TT1");
+        jmri.Transit transit = new jmri.implementation.DefaultTransit("TT1");
         ActiveTrain t = new ActiveTrain(transit,"Train",ActiveTrain.USER);
         Assert.assertNotNull("exists",t);
     }
