+++ java/test/jmri/jmrit/dispatcher/AllocatedSectionTest.java
@@ -20,7 +20,7 @@ public class AllocatedSectionTest {
         Assume.assumeFalse(GraphicsEnvironment.isHeadless());
         OptionsFile.setDefaultFileName("java/test/jmri/jmrit/dispatcher/dispatcheroptions.xml");  // exist?
         DispatcherFrame d = InstanceManager.getDefault(DispatcherFrame.class);
-        jmri.Transit transit = new jmri.Transit("TT1");
+        jmri.Transit transit = new jmri.implementation.DefaultTransit("TT1");
         ActiveTrain at = new ActiveTrain(transit, "Train", ActiveTrain.USER);
         jmri.Section section1 = new jmri.implementation.DefaultSection("TS1");
         jmri.Section section2 = new jmri.implementation.DefaultSection("TS2");
